# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CapsRefCoverageReport": {
        "additionalProperties": false,
        "properties": {
          "caps_ref": {
            "title": "Caps Ref",
            "type": "string"
          },
          "layers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CoverageLayerCounts"
            },
            "title": "Layers",
            "type": "object"
          },
          "scope_id": {
            "title": "Scope Id",
            "type": "string"
          }
        },
        "required": [
          "scope_id",
          "caps_ref",
          "layers"
        ],
        "title": "CapsRefCoverageReport",
        "type": "object"
      },
...
      "ContentArtifactProvenanceResponse": {
        "properties": {
          "artifact_hash": {
            "title": "Artifact Hash",
            "type": "string"
          },
          "artifact_id": {
            "format": "uuid",
            "title": "Artifact Id",
            "type": "string"
          },
          "source_snapshot_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Snapshot Hash"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ContentArtifactProvenanceSourceResponse"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "artifact_id",
          "status",
          "artifact_hash",
          "source_snapshot_hash",
          "sources"
        ],
        "title": "ContentArtifactProvenanceResponse",
        "type": "object"
      },
...
      "ContentArtifactValidationRequest": {
        "properties": {
          "artifact_json": {
            "title": "Artifact Json",
            "type": "object"
          },
          "artifact_type": {
            "$ref": "#/components/schemas/ContentArtifactType"
          },
          "caps_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caps Ref"
          },
          "min_sources": {
            "default": 1,
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Min Sources",
            "type": "integer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ETLSourceCitation"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "artifact_type",
          "artifact_json"
        ],
        "title": "ContentArtifactValidationRequest",
        "type": "object"
      },
...
      "ContentFactoryETLStatusResponse": {
        "properties": {
          "mcp_runtime_imported": {
            "title": "Mcp Runtime Imported",
            "type": "boolean"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "pipeline_package": {
            "title": "Pipeline Package",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "pipeline_package",
          "mcp_runtime_imported"
        ],
        "title": "ContentFactoryETLStatusResponse",
        "type": "object"
      },
...
      "ContentLayer": {
        "enum": [
          "topic_map",
          "diagnostic_items",
          "lessons",
          "assessment_blueprints",
          "study_plan_templates"
        ],
        "title": "ContentLayer",
        "type": "string"
      },
      "ContentScope": {
        "additionalProperties": false,
        "properties": {
          "caps_refs": {
            "items": {
              "type": "string"
            },
            "title": "Caps Refs",
            "type": "array"
          },
          "curriculum": {
            "minLength": 1,
            "title": "Curriculum",
            "type": "string"
          },
          "grade": {
            "maximum": 12.0,
            "minimum": 0.0,
            "title": "Grade",
            "type": "integer"
          },
          "language": {
            "maxLength": 8,
            "minLength": 2,
            "title": "Language",
            "type": "string"
          },
          "scope_id": {
            "minLength": 1,
            "title": "Scope Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ContentScopeStatus"
          },
          "subject": {
            "minLength": 1,
            "title": "Subject",
            "type": "string"
          },
          "subject_code": {
            "minLength": 1,
            "title": "Subject Code",
            "type": "string"
          },
          "topic_map_path": {
            "minLength": 1,
            "title": "Topic Map Path",
            "type": "string"
          }
        },
        "required": [
          "scope_id",
          "grade",
          "subject_code",
          "subject",
          "language",
          "curriculum",
          "status",
          "topic_map_path"
        ],
        "title": "ContentScope",
        "type": "object"
      },
      "ContentScopeStatus": {
        "enum": [
          "draft",
          "active",
          "deprecated",
          "archived"
        ],
        "title": "ContentScopeStatus",
        "type": "string"
      },
...
      "CoverageTarget": {
        "additionalProperties": false,
        "properties": {
          "caps_ref": {
            "minLength": 1,
            "title": "Caps Ref",
            "type": "string"
          },
          "scope_id": {
            "minLength": 1,
            "title": "Scope Id",
            "type": "string"
          },
          "targets": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Targets",
            "type": "object"
          }
        },
        "required": [
          "scope_id",
          "caps_ref"
        ],
        "title": "CoverageTarget",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScopeCoverageReport": {
        "additionalProperties": false,
        "properties": {
          "grade": {
            "title": "Grade",
            "type": "integer"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "layers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ScopeCoverageLayerSummary"
            },
            "title": "Layers",
            "type": "object"
          },
          "per_caps_ref": {
            "items": {
              "$ref": "#/components/schemas/CapsRefCoverageReport"
            },
            "title": "Per Caps Ref",
            "type": "array"
          },
          "scope_id": {
            "title": "Scope Id",
            "type": "string"
          },
          "subject_code": {
            "title": "Subject Code",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/ScopeCoverageSummary"
          }
        },
        "required": [
          "scope_id",
          "grade",
          "subject_code",
          "language",
          "summary",
          "layers",
          "per_caps_ref"
        ],
        "title": "ScopeCoverageReport",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered adaptive learning platform — Grade R to 7. CAPS-aligned. POPIA-compliant.",
    "title": "EduBoost SA V2",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/admin/content-factory/etl/status": {
      "get": {
        "operationId": "etl_status_api_v2_admin_content_factory_etl_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentFactoryETLStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Etl Status",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
...
    "/api/v2/admin/content-factory/provenance/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_provenance_api_v2_admin_content_factory_provenance__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentArtifactProvenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Artifact Provenance",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/api/v2/admin/content-factory/scopes": {
      "get": {
        "operationId": "list_content_scopes_api_v2_admin_content_factory_scopes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContentScope"
                  },
                  "title": "Response List Content Scopes Api V2 Admin Content Factory Scopes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Content Scopes",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/api/v2/admin/content-factory/scopes/{scope_id}": {
      "get": {
        "operationId": "get_content_scope_api_v2_admin_content_factory_scopes__scope_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentScope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Content Scope",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/api/v2/admin/content-factory/scopes/{scope_id}/coverage": {
      "get": {
        "operationId": "get_content_scope_coverage_api_v2_admin_content_factory_scopes__scope_id__coverage_get",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "layer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ContentLayer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeCoverageReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Content Scope Coverage",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/api/v2/admin/content-factory/scopes/{scope_id}/coverage/{caps_ref}": {
      "get": {
        "operationId": "get_content_caps_ref_coverage_api_v2_admin_content_factory_scopes__scope_id__coverage__caps_ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "caps_ref",
            "required": true,
            "schema": {
              "title": "Caps Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "layer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ContentLayer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsRefCoverageReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Content Caps Ref Coverage",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/api/v2/admin/content-factory/scopes/{scope_id}/targets": {
      "get": {
        "operationId": "get_content_scope_targets_api_v2_admin_content_factory_scopes__scope_id__targets_get",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CoverageTarget"
                  },
                  "title": "Response Get Content Scope Targets Api V2 Admin Content Factory Scopes  Scope Id  Targets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Content Scope Targets",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/api/v2/admin/content-factory/validate-artifact": {
      "post": {
        "operationId": "validate_artifact_payload_api_v2_admin_content_factory_validate_artifact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentArtifactValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentArtifactValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate Artifact Payload",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
...
    "/v2/admin/content-factory/etl/status": {
      "get": {
        "operationId": "etl_status_v2_admin_content_factory_etl_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentFactoryETLStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Etl Status",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
...
    "/v2/admin/content-factory/provenance/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_provenance_v2_admin_content_factory_provenance__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentArtifactProvenanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Artifact Provenance",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/v2/admin/content-factory/scopes": {
      "get": {
        "operationId": "list_content_scopes_v2_admin_content_factory_scopes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContentScope"
                  },
                  "title": "Response List Content Scopes V2 Admin Content Factory Scopes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Content Scopes",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/v2/admin/content-factory/scopes/{scope_id}": {
      "get": {
        "operationId": "get_content_scope_v2_admin_content_factory_scopes__scope_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentScope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Content Scope",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/v2/admin/content-factory/scopes/{scope_id}/coverage": {
      "get": {
        "operationId": "get_content_scope_coverage_v2_admin_content_factory_scopes__scope_id__coverage_get",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "layer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ContentLayer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeCoverageReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Content Scope Coverage",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/v2/admin/content-factory/scopes/{scope_id}/coverage/{caps_ref}": {
      "get": {
        "operationId": "get_content_caps_ref_coverage_v2_admin_content_factory_scopes__scope_id__coverage__caps_ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "caps_ref",
            "required": true,
            "schema": {
              "title": "Caps Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "layer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ContentLayer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Layer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapsRefCoverageReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Content Caps Ref Coverage",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/v2/admin/content-factory/scopes/{scope_id}/targets": {
      "get": {
        "operationId": "get_content_scope_targets_v2_admin_content_factory_scopes__scope_id__targets_get",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CoverageTarget"
                  },
                  "title": "Response Get Content Scope Targets V2 Admin Content Factory Scopes  Scope Id  Targets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Content Scope Targets",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
    "/v2/admin/content-factory/validate-artifact": {
      "post": {
        "operationId": "validate_artifact_payload_v2_admin_content_factory_validate_artifact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentArtifactValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentArtifactValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate Artifact Payload",
        "tags": [
          "admin-content-factory"
        ]
      }
    },
...
  },
...
}